requirements.txt
requests>=2.28
pandas>=2.0
pyarrow
python-dotenv
tqdm